src/python/software/software_v1/requirements.txt
python-statemachine
pandas
numpy
matplotlib
scipy
PySide6
pyserial